Update WebSockets.cpp to eliminate inclusion warning/wrapper.
Hi, Dave/Rutger.

I know we don't like fixing forks, but upstream is abandoned. Confession: This is a web edit of an "obvious" one-liner. Treat with skepticism/flippancy as you see fit.


Fix compiler warning on ESP-IDF 3. We use ESP-IDF5. IDF3 was new in 2018.

I tracked this back to 2019, where I pretty much ran out of interest. https://github.com/espressif/esp-idf/issues/3245

This reduces warnings in (almost?) all NightDriverLED builds and fixes a build error (five years of deprecation warnings isn't UNreasonable) in an upcoming ESP-IDF that we don't like to talk about yet.

Testing things like this in PlatformIO is gnarly, but I locally edited .pio/libdeps/*/RemoteDebug/src/utility/WebSockets.cpp, ran a buddybuild, and confirmed that Mesmerizer still worked. The change is dead obvious; it was just a rename, and this eliminates the old version that hocks up a #warning and includes the new version, so this shouldn't be too exciting.

RemoteDebug was apparently abandoned in 2019. This is pain relief until I (or someone... please) can perform the major surgery needed for newer SOCs. Future direction currently unclear.

Still, it's better than what we have for now and shouldn't be TOO contentious.